fix(v1.1.76): compile errors in desktop printing
- windows_printer: import package:ffi/ffi.dart (calloc lives there,
  not in dart:ffi) + declare ffi ^2.1.3 in pubspec. Fixes the 17
  'calloc not found' errors on both APK & Windows builds.
- image_pick: pickImageBytes maxWidth is double to match
  image_picker's pickImage(maxWidth:) signature.
- scripts/github mirror: failure-annotation diagnostics steps
  (staged for the next owner apply). [fix]

+++ flutter_app/lib/ui/widgets/image_pick.dart
@@ -12,7 +12,7 @@ import '../../core/platform_check.dart';
 /// image_picker backend, so we go through the OS file dialog (file_picker)
 /// instead — same bytes either way.
 Future<List<int>?> pickImageBytes({
-  int maxWidth = 600,
+  double maxWidth = 600,
   int imageQuality = 70,
 }) async {
   if (!kIsWeb && OfPlatform.isDesktop) {
